+++ src/fb_pdnstools/xlate.py
@@ -3,8 +3,7 @@
 """
 @summary: The module for i18n.
 
-          It provides translation object, usable from all other
-          modules in this package.
+          It provides translation object, usable from all other modules in this package.
 @author: Frank Brehm
 @contact: frank@brehm-online.com
 @copyright: © 2024 by Frank Brehm, Berlin
@@ -16,46 +15,59 @@ import copy
 import gettext
 import logging
 import sys
-try:
-    from pathlib import Path
-except ImportError:
-    from pathlib2 import Path
-try:
-    from packaging.version import Version
-except ImportError:
-    from distutils.version import LooseVersion as Version
+from pathlib import Path
 
 # Third party modules
 import babel
 import babel.lists
 from babel.support import Translations
 
+try:
+    from semver import Version
+except ImportError:
+    from semver import VersionInfo as Version
+
 DOMAIN = 'fb_pdnstools'
 
 LOG = logging.getLogger(__name__)
 
-__version__ = '0.1.2'
+__version__ = '0.2.0'
 
 __me__ = Path(__file__).resolve()
 __module_dir__ = __me__.parent
 __lib_dir__ = __module_dir__.parent
 __base_dir__ = __lib_dir__.parent
-LOCALE_DIR = __base_dir__.joinpath('locale')
+
+LOCALE_DIR = __base_dir__ / "locale"
+
 if LOCALE_DIR.is_dir():
+    # Not installed, in development workdir
     LOCALE_DIR = str(LOCALE_DIR)
 else:
+    # Somehow installed
     LOCALE_DIR = __module_dir__.joinpath('locale')
-    if LOCALE_DIR.is_dir():
-        LOCALE_DIR = str(LOCALE_DIR)
+    if sys.prefix == sys.base_prefix:
+        # installed as a package
+        LOCALE_DIR = sys.prefix + "/share/locale"
     else:
-        LOCALE_DIR = sys.prefix + '/share/locale'
+        # Obviously in a virtual environment
+        __base_dir__ = Path(sys.prefix)
+        LOCALE_DIR = __base_dir__ / "share" / "locale"
+        if LOCALE_DIR.is_dir():
+            LOCALE_DIR = str(LOCALE_DIR.resolve())
+        else:
+            LOCALE_DIR = __module_dir__ / "locale"
+            if LOCALE_DIR.is_dir():
+                LOCALE_DIR = str(LOCALE_DIR)
+            else:
+                LOCALE_DIR = str(__base_dir__ / sys.prefix / "share" / "locale")
 
 DEFAULT_LOCALE_DEF = 'en_US'
 DEFAULT_LOCALE = babel.core.default_locale()
 if not DEFAULT_LOCALE:
     DEFAULT_LOCALE = DEFAULT_LOCALE_DEF
 
-__mo_file__ = gettext.find(DOMAIN, str(LOCALE_DIR))
+__mo_file__ = gettext.find(DOMAIN, LOCALE_DIR)
 if __mo_file__:
     try:
         with open(__mo_file__, 'rb') as F:
@@ -65,13 +77,10 @@ if __mo_file__:
 else:
     XLATOR = gettext.NullTranslations()
 
-CUR_BABEL_VERSION = Version(babel.__version__)
-NEWER_BABEL_VERSION = Version('2.6.0')
+CUR_BABEL_VERSION = Version.parse(babel.__version__)
+NEWER_BABEL_VERSION = Version.parse('2.6.0')
 
-SUPPORTED_LANGS = (
-    'de',
-    'en'
-)
+SUPPORTED_LANGS = ('de', 'en')
 
 _ = XLATOR.gettext
 
@@ -102,11 +111,23 @@ def format_list(lst, do_repr=False, style='standard', locale=DEFAULT_LOCALE):
 
 if __name__ == '__main__':
 
-    print(_('Module directory: {!r}').format(__module_dir__))
-    print(_('Base directory: {!r}').format(__base_dir__))
-    print(_('Locale directory: {!r}').format(LOCALE_DIR))
-    print(_('Locale domain: {!r}').format(DOMAIN))
-    print(_('Found .mo-file: {!r}').format(__mo_file__))
+    out_list = []
+    out_list.append([_("Module directory:"), str(__module_dir__)])
+    out_list.append([_("Lib directory:"), str(__lib_dir__)])
+    out_list.append([_("Base directory:"), str(__base_dir__)])
+    out_list.append([_("Locale directory:"), LOCALE_DIR])
+    out_list.append([_("Locale domain:"), DOMAIN])
+    out_list.append([_("Found .mo-file:"), __mo_file__])
+
+    max_len = 1
+    for pair in out_list:
+        if len(pair[0]) > max_len:
+            max_len = len(pair[0])
+
+    template = "{{label:<{}}} {{val!r}}".format(max_len)
+    for pair in out_list:
+        print(template.format(label=pair[0], val=pair[1]))
+
 
 # =============================================================================