+++ CHANGELOG.md
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Cleaning up no more used stuff.
 * Updating `update-env.sh` for refactored virtual env.
 
+### Fixed
+
+*  Fixing `src/fb_vmware/xlate.py`.
+
 ## [1.0.2] - 2024-07-01
 
 ### Added